code-library/misc-13/requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.19.0

# Scientific Computing
scipy>=1.7.0
scikit-learn>=1.0.0

# Deep Learning Utilities
einops>=0.6.0

# Visualization
matplotlib>=3.5.0
PyQt5>=5.15.0

# Configuration
PyYAML>=6.0
easydict>=1.9.0

# Experiment Tracking
wandb>=0.12.0

# Optional: For distributed training
# horovod>=0.25.0  # Uncomment if using distributed training